+++ src/metered_concurrency.rs
@@ -22,7 +22,7 @@ pub struct CountingSemaphoreGuard<'a> {
 impl CountingSemaphore {
     // Create a semaphore with `max` available resources and a linearly increasing backoff of
     // `backoff` (used during spinlock contention).
-    pub fn new(max: uint, backoff: Duration) -> CountingSemaphore {
+    pub fn new(max: usize, backoff: Duration) -> CountingSemaphore {
         CountingSemaphore { count: AtomicUint::new(max), backoff: backoff }
     }
 
@@ -48,7 +48,7 @@ impl CountingSemaphore {
     }
 
     // Return remaining resource count
-    pub fn count(&self) -> uint {
+    pub fn count(&self) -> usize {
         self.count.load(Ordering::SeqCst)
     }
 }
@@ -62,7 +62,7 @@ impl<'a> Drop for CountingSemaphoreGuard<'a> {
 }
 
 fn metered(duration: Duration) {
-    static MAX_COUNT: uint = 4; // Total available resources
+    static MAX_COUNT: usize = 4; // Total available resources
     static NUM_WORKERS: u8 = 10; // Number of workers contending for the resources
     let backoff = Duration::milliseconds(1); // Linear backoff time
     // Create a shared reference to the semaphore
